test: add tests for missing config file or section.
